chore: 폴더 구조 수정 및 파일 경로 설정 (#3)
